+++ src/main/java/com/rpgvtt/montador_de_rpg_backend/domain/model/Resolucao.java
@@ -5,6 +5,9 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import org.hibernate.annotations.JdbcTypeCode;
+import org.hibernate.type.SqlTypes;
+import tools.jackson.databind.JsonNode;
 
 import java.util.List;
 
@@ -16,10 +19,17 @@ import java.util.List;
 @Table(name = "Resolucao")
 public class Resolucao {
     @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "evento_sistema_seq")
-    @SequenceGenerator(name = "evento_sistema_seq", sequenceName = "evento_sistema_sequence", allocationSize = 1)
+    @GeneratedValue(
+            strategy = GenerationType.SEQUENCE,
+            generator = "evento_sistema_seq"
+    )
+    @SequenceGenerator(
+            name = "evento_sistema_seq",
+            sequenceName = "evento_sistema_sequence",
+            allocationSize = 1
+    )
     @Column(name = "id_resolucao")
-    private Integer idResolucao;
+    private Long idResolucao;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "Rolagem_id_rolagem")
